// src/app/core/models/order.model.ts
import { Product } from './product.model';

export interface OrderItem extends Partial<Product> {
  quantity: number;
}

export interface OrderAddress {
  street: string;
  number: string;
  city: string;
}

export interface Order {
  id: string;
  userId: string;
  user?: { name: string; email: string };
  items: OrderItem[];
  address: OrderAddress;
  total: number;
  createdAt: number;
}